# IO: Async disk I/O on Windows