--- a/src/tigerbeetle/main.zig
+++ b/src/tigerbeetle/main.zig
@@ -172,8 +172,12 @@ const Command = struct {
         else
             .direct_io_required;
 
+        command.io = try IO.init(128, 0);
+        errdefer command.io.deinit();
+
         const basename = std.fs.path.basename(path);
         command.fd = try IO.open_file(
+            &command.io,
             command.dir_fd,
             basename,
             data_file_size_min,
@@ -181,9 +185,6 @@ const Command = struct {
             direct_io,
         );
         errdefer os.close(command.fd);
-
-        command.io = try IO.init(128, 0);
-        errdefer command.io.deinit();
 
         command.storage = try Storage.init(&command.io, command.fd);
         errdefer command.storage.deinit();