@batiati 2024-05-10
Wow, it's actually faster than the Linux version, isn't it?